split quickstart from other assistant tests
Saves 20-30s (~15%) in total test runtime versus the parent.

CHANGELOG_BEGIN
CHANGELOG_END
